requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
scipy>=1.10.0
requests>=2.31.0
streamlit>=1.28.0
Pillow>=10.0.0
tqdm>=4.65.0
